fixed form options and more refactoring
